moved Fake* classes to fast_carpenter.testing
